Refactor hdsfhir [stage 1]: Move everything to models package
