+++ models/core.go
@@ -1,4 +1,4 @@
-package hdsfhir
+package models
 
 import (
 	"time"
