+++ models/core_uploadable.go
@@ -1,4 +1,4 @@
-package hdsfhir
+package models
 
 import (
 	"bytes"
