+++ models/medication_test.go
@@ -1,4 +1,4 @@
-package hdsfhir
+package models
 
 import (
 	"encoding/json"
@@ -15,7 +15,7 @@ type MedicationSuite struct {
 var _ = Suite(&EncounterSuite{})
 
 func (s *MedicationSuite) SetUpSuite(c *C) {
-	data, err := ioutil.ReadFile("./fixtures/john_peters.json")
+	data, err := ioutil.ReadFile("../fixtures/john_peters.json")
 	util.CheckErr(err)
 	patient := &Patient{}
 	err = json.Unmarshal(data, patient)
